feat: now can input the keyword by interest, and paragraph related to that topic generates
