+++ ias/views/base_views.py
@@ -13,9 +13,6 @@ def ai_index(request):
         ai_list = ai_list.filter(
             Q(subject__icontains=kw) | # 제목 검색
             Q(content__icontains=kw)   # 내용 검색
-            # Q(input__content__icontains=kw) |  # 답변 내용 검색
-            # Q(author__username__icontains=kw) |  # 질문 글쓴이 검색
-            # Q(input__author__username__icontains=kw)  # 답변 글쓴이 검색
         ).distinct()
     paginator = Paginator(ai_list, 10) # 페이지당 10개씩 보여주기
     page_obj = paginator.get_page(page)
