+++ templates/ias/ai_list.html
@@ -5,12 +5,22 @@
     <div class="row my-3">
 <!--        <div class="col-6">-->
 <!--        </div>-->
-        <div class="col-6">
-            <a href="{% url 'ias:ai_create' %}" class="btn btn-primary">Generate by AI</a>
-        </div>
-        <div class="col-6">
+        <div class="col-md-6">
             <div class="input-group">
-                <input type="text" id="search_kw" class="form-control" value="{{ kw|default_if_none:'' }}">
+                <form id="topicForm" method="get" action="{% url 'ias:ai_create' %}">
+                {% csrf_token %}
+                <div class="input-group">
+                    <input type="text" id="topic_input" name="topic" class="form-control" placeholder="Enter Topic">
+                    <div class="input-group-append">
+                        <button class="btn btn-outline-primary" type="submit">Generate by AI</button>
+                    </div>
+                </div>
+            </form>
+
+            </div>
+            <!-- Second input field -->
+            <div class="input-group mt-3">
+                <input type="text" id="search_kw" class="form-control" placeholder="Search...">
                 <div class="input-group-append">
                     <button class="btn btn-outline-secondary" type="button" id="btn_search">Search</button>
                 </div>
@@ -36,9 +46,9 @@
             </td>
             <td class="text-start">
                 <a href="{% url 'ias:ai_detail' ai.id %}">{{ ai.subject }}</a>
-<!--                {% if ai.input_set.count > 0 %}-->
-<!--                <span class="text-danger small mx-2">{{ ai.input_set.count }}</span>-->
-<!--                {% endif %}-->
+                {% if ai.input_set.count > 0 %}
+                <span class="text-danger small mx-2">{{ ai.input_set.count }}</span>
+                {% endif %}
             </td>
 <!--            <td>{{ ai.author.username}}</td> &lt;!&ndash; 글쓴이 추가 &ndash;&gt;-->
             <td>{{ ai.create_date }}</td>
@@ -95,10 +105,7 @@
         {% endif %}
     </ul>
 </div>
-<form id="searchForm" method="get" action="{% url 'ai_index' %}">
-    <input type="hidden" id="kw" name="kw" value="{{ kw|default_if_none:'' }}">
-    <input type="hidden" id="page" name="page" value="{{ page }}">
-</form>
+
 {% endblock %}
 {% block script %}
 <script type='text/javascript'>
